Refactor src/ structure
