+++ src/pages/_app.tsx
@@ -1,7 +1,7 @@
 import '../styles/globals.css';
 import type { AppProps } from 'next/app';
 import Head from 'next/head';
-import Header from '../src/components/Header/Header';
+import Header from '../components/Header/Header';
 
 function MyApp({ Component, pageProps }: AppProps) {
   return (
